Move dataloaders to appropriate module
