+++ helix/urls.py
@@ -19,37 +19,11 @@ from django.contrib import admin
 from django.conf import settings
 from django.urls import path, re_path, include
 from django.views.decorators.csrf import csrf_exempt
-from django.utils.functional import cached_property
 # from graphene_django.views import GraphQLView
 from graphene_file_upload.django import FileUploadGraphQLView
-from apps.country.dataloaders import OneToManyLoader, CountLoader
 
 from . import api_urls as rest_urls
-
-
-class GQLContext:
-    def __init__(self, request):
-        self.request = request
-        self.one_to_many_dataloaders = {}
-        self.count_dataloaders = {}
-
-    @cached_property
-    def user(self):
-        return self.request.user
-
-    def get_dataloader(self, parent: str, child: str):
-        # TODO: rename to get OneToManyLoader?
-        # return a different dataloader for each ref
-        ref = f'{parent}_{child}'
-        if ref not in self.one_to_many_dataloaders:
-            self.one_to_many_dataloaders[ref] = OneToManyLoader()
-        return self.one_to_many_dataloaders[ref]
-
-    def get_count_loader(self, parent: str, child: str):
-        ref = f'{parent}_{child}'
-        if ref not in self.count_dataloaders:
-            self.count_dataloaders[ref] = CountLoader()
-        return self.count_dataloaders[ref]
+from utils.graphene.context import GQLContext
 
 
 class CustomGraphQLView(FileUploadGraphQLView):
